fix: reset api
